Cleaned up .multiSelectionResponsive code and documentation.
The 'dims' argument now only accepts a single dimension, either row or column;
this is possible as the entire .multiSelectionResponsive generic is only called
in one place with a non-"none" output from .multiSelectionDimension(). It
allows us to greatly simplify the implementation of various methods.

Updated the documentation and signature to mention the new expectations. Also
added a default Panel method that conservatively requires a re-render upon any
multiple selection; subclasses can refine this for efficiency.